requirements.txt
streamlit==1.22.0
pandas==2.0.1
geopandas==0.13.0
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
python-dotenv==1.0.0
folium==0.14.0
scikit-learn==1.2.2
numpy==1.24.3
pytest==7.3.1
alembic==1.11.1  # for database migrations
flask==2.0.1
flask-sqlalchemy==2.5.1

# Floor plan scraping dependencies
selenium==4.16.0
pillow==10.1.0
webdriver-manager==4.0.1  # Automatic chromedriver management